edit news

+++ src/scripts/profile/newsEventListener.js
@@ -1,26 +1,27 @@
 import newsForm from "./newsForm"
 import printToDom from "../utilities/printToDOM"
-import API from "../utilities/apiManager";
-import createNews from "./createNews";
+import API from "../utilities/apiManager"
+import createNews from "./createNews"
+import editNewsForm from "./editNewsForm"
 
 const newsEventListener = () => {
     document.querySelector("#news").addEventListener("click", (event) => {
+        let userId = document.querySelector("#userId").value
+        let buttonParent = event.target.parentElement
+
         // ADD NEWS ITEM button
         if (event.target.id === "addNews") {
             let newsFormHTML = newsForm()
             printToDom(newsFormHTML, "#newsFormSection")
 
-            // SUBMIT button
+        // SUBMIT button
         } else if (event.target.id === "submitNews") {
             let news = document.querySelector("#newsTitle").value
-            let date = document.querySelector("#newsDate").value
             let newsSynopsis = document.querySelector("#newsSynopsis").value
             let newsURL = document.querySelector("#newsURL").value
-            let userId = document.querySelector("#userId").value
 
             let newNewsObject = {
                 news: news,
-                date: date,
                 newsSynopsis: newsSynopsis,
                 newsURL: newsURL,
                 userId: parseInt(userId)
@@ -31,19 +32,55 @@ const newsEventListener = () => {
 
         // DELETE button
         } else if (event.target.id.startsWith("deleteNews")) {
-            // Find ID of news item to remove
-            let idToDelete = event.target.id.split("--")[1]
-            let stringToDelete = `news/${idToDelete}`
-            // Delete from API
-            API.DELETE(stringToDelete)
-                // Re-load news list
-                .then(response => {
-                    createNews()
-                })
+
+            // Check that news belongs to active user
+            if (buttonParent.classList.contains(`user--${userId}`)) {
+                // Find ID of news item to remove
+                let idToDelete = event.target.id.split("--")[1]
+                let stringToDelete = `news/${idToDelete}`
+                // Delete from API
+                API.DELETE(stringToDelete)
+                    // Re-load news list
+                    .then(response => {
+                        createNews()
+                    })
+            }
 
         // EDIT button
-        } else if (event.target.id.startsWith("editNews")) {
+        } else if (event.target.id.startsWith("editNewsButton")) {
 
+            // Check that news belongs to active user
+            if (buttonParent.classList.contains(`user--${userId}`)) {
+                // Find ID of news item to edit
+                let idToEdit = event.target.id.split("--")[1]
+                // Show edit form
+                let editFormHTML = editNewsForm(idToEdit)
+                printToDom(editFormHTML, "#newsFormSection")
+
+                // UPDATE button
+                document.querySelector("#updateNews").addEventListener("click", () => {
+                    let news = document.querySelector("#editNewsTitle").value
+                    let newsSynopsis = document.querySelector("#editNewsSynopsis").value
+                    let newsURL = document.querySelector(`#newsLink--${idToEdit}`).href
+
+                    let updatedNewsObject = {
+                        news: news,
+                        newsSynopsis: newsSynopsis,
+                        newsURL: newsURL,
+                        userId: parseInt(userId)
+                    }
+
+                    let apiString = `news/${idToEdit}`
+                    // API call
+                    API.EDIT(apiString, updatedNewsObject)
+                        .then(response => {
+                            createNews()
+                        })
+
+                    // Remove edit form
+                    document.querySelector("#newsFormSection").innerHTML = ""
+                })
+            }
         }
     })
 }
